backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg[binary]==3.1.18
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
boto3==1.34.22
requests==2.31.0
sentence-transformers==2.3.1
qdrant-client==1.7.0
redis==5.0.1
numpy==1.26.3
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
PyPDF2==3.0.1
python-docx==1.1.0
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2
